backend/genstoryai_backend/pyproject.toml
[tool.poetry]
name = "genstoryai_backend"
version = "0.1.0"
description = "GenStoryAI backend,fastapi server"
authors = ["fengchuimailang <1050575425@qq.com>"]
readme = "README.md"
packages = [{include = "genstoryai_backend", from = "."}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
flake8 = ">=7.3.0,<8.0.0"
pytest = ">=8.4.1,<9.0.0"
fastapi = ">=0.115.14,<0.116.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic = ">=2.11.7,<3.0.0"
pydantic-ai = ">=0.3.5,<0.4.0"
langfuse = ">=3.0.8,<4.0.0"
sqlmodel = ">=0.0.8,<0.1.0"
mysqlclient = ">=2.2.0,<3.0.0"
python-dotenv = ">=1.0.0,<2.0.0"

[tool.poetry.scripts]
start = "genstoryai_backend.run:main"
test = "pytest:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
